Move tests to correct file and use injected ssh_connection instead of setting it up from scratch
